refactor : repository layer 제거
