+++ src/main/java/racingcar/service/RacingGameService.java
@@ -3,16 +3,21 @@ package racingcar.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.stereotype.Service;
+import racingcar.dao.GameDao;
+import racingcar.dao.PlayerDao;
 import racingcar.domain.Car;
 import racingcar.domain.CarGenerator;
 import racingcar.domain.RacingGame;
 import racingcar.domain.Winner;
 import racingcar.dto.PlayResultResponseDto;
-import racingcar.repository.RacingGameRepository;
+import racingcar.dto.WinnerFormatter;
 import racingcar.utils.NumberGenerator;
 import racingcar.utils.RandomNumberGenerator;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
+import java.util.Locale;
 
 @Service
 public class RacingGameService {
@@ -23,14 +28,16 @@ public class RacingGameService {
     @Autowired
     ConversionService conversionService;
     @Autowired
-    RacingGameRepository racingGameRepository;
+    private GameDao gameDao;
+    @Autowired
+    private PlayerDao playerDao;
 
     public PlayResultResponseDto run(String names, Integer count) {
         RacingGame racingGame = play(names, count);
         Winner winner = racingGame.getWinner();
         List<Car> cars = racingGame.getCars();
 
-        racingGameRepository.insert(winner, count, cars);
+        insert(winner, count, cars);
 
         return new PlayResultResponseDto(winner, cars);
     }
@@ -44,7 +51,23 @@ public class RacingGameService {
 
     }
 
-    public List<PlayResultResponseDto> get() {
-        return racingGameRepository.get();
+    public void insert(Winner winner, int count, List<Car> cars) {
+        String winnerNames = new WinnerFormatter().print(winner, Locale.getDefault());
+        int gameId = gameDao.insert(winnerNames, count);
+        playerDao.insert(gameId, cars);
+    }
+
+    public List<PlayResultResponseDto> getPlayHistory() {
+        List<PlayResultResponseDto> playResultResponseDtos = new ArrayList<>();
+
+        int lastGameId = gameDao.findLastId();
+        for (int index = 1; index < lastGameId + 1; index++) {
+            String winnerNames = gameDao.findWinners(index);
+            Winner winner = new Winner(Arrays.asList(winnerNames.split(",")));
+            List<Car> cars = playerDao.find(index);
+
+            playResultResponseDtos.add(new PlayResultResponseDto(winner, cars));
+        }
+        return playResultResponseDtos;
     }
 }
