+++ src/main/java/racingcar/view/Input.java
@@ -3,7 +3,6 @@ package racingcar.view;
 import java.util.Scanner;
 
 public class Input {
-    private static final String DELIMITER = ",";
     private static final Validator validator = new Validator();
     private static final Scanner scanner = new Scanner(System.in);
